+++ src/drivers/shadow/shadow.c
@@ -69,7 +69,7 @@ static int vshadow_init(struct vxpdb_state *vp, const char *config_file)
 	return 1;
 }
 
-static int vshadow_open(struct vxpdb_state *vp, long flags)
+static int vshadow_open(struct vxpdb_state *vp, unsigned int flags)
 {
 	struct shadow_state *state = vp->state;
 	int ret;
