+++ src/libvxpdb/dummy.c
@@ -21,7 +21,7 @@ static int vdummy_init(struct vxpdb_state *vp, const char *config_file)
 	return 1;
 }
 
-static int vdummy_open(struct vxpdb_state *vp, long flags)
+static int vdummy_open(struct vxpdb_state *vp, unsigned int flags)
 {
 	return 1;
 }
